+++ library/src/android/support/v4/app/FragmentActivity.java
@@ -1050,16 +1050,6 @@ public class FragmentActivity extends Activity {
         return actionMode;
     }
 
-    /**
-     * Get a special instance of {@link MenuItemImpl} which denotes the home
-     * item and should be invoked when the custom home button is clicked.
-     *
-     * @return Menu item instance.
-     */
-    final MenuItemImpl getHomeMenuItem() {
-        return mSupportMenu.addDetached(android.R.id.home);
-    }
-
     // ------------------------------------------------------------------------
     // FRAGMENT SUPPORT
     // ------------------------------------------------------------------------